+++ oci/spec_opts_unix.go
@@ -20,21 +20,15 @@ package oci
 
 import (
 	"context"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 
 	"github.com/containerd/containerd/containers"
-	specs "github.com/opencontainers/runtime-spec/specs-go"
-	"github.com/pkg/errors"
-	"golang.org/x/sys/unix"
 )
 
 // WithHostDevices adds all the hosts device nodes to the container's spec
 func WithHostDevices(_ context.Context, _ Client, _ *containers.Container, s *Spec) error {
 	setLinux(s)
 
-	devs, err := getDevices("/dev", "")
+	devs, err := HostDevices()
 	if err != nil {
 		return err
 	}
@@ -42,11 +36,9 @@ func WithHostDevices(_ context.Context, _ Client, _ *containers.Container, s *Sp
 	return nil
 }
 
-var errNotADevice = errors.New("not a device node")
-
 // WithDevices recursively adds devices from the passed in path and associated cgroup rules for that device.
 // If devicePath is a dir it traverses the dir to add all devices in that dir.
-// If devicePath is not a dir, it attempts to add the signle device.
+// If devicePath is not a dir, it attempts to add the single device.
 func WithDevices(devicePath, containerPath, permissions string) SpecOpts {
 	return func(_ context.Context, _ Client, _ *containers.Container, s *Spec) error {
 		devs, err := getDevices(devicePath, containerPath)
@@ -56,107 +48,6 @@ func WithDevices(devicePath, containerPath, permissions string) SpecOpts {
 		s.Linux.Devices = append(s.Linux.Devices, devs...)
 		return nil
 	}
-}
-
-func getDevices(path, containerPath string) ([]specs.LinuxDevice, error) {
-	stat, err := os.Stat(path)
-	if err != nil {
-		return nil, errors.Wrap(err, "error stating device path")
-	}
-
-	if !stat.IsDir() {
-		dev, err := deviceFromPath(path)
-		if err != nil {
-			return nil, err
-		}
-		if containerPath != "" {
-			dev.Path = containerPath
-		}
-		return []specs.LinuxDevice{*dev}, nil
-	}
-
-	files, err := ioutil.ReadDir(path)
-	if err != nil {
-		return nil, err
-	}
-	var out []specs.LinuxDevice
-	for _, f := range files {
-		switch {
-		case f.IsDir():
-			switch f.Name() {
-			// ".lxc" & ".lxd-mounts" added to address https://github.com/lxc/lxd/issues/2825
-			// ".udev" added to address https://github.com/opencontainers/runc/issues/2093
-			case "pts", "shm", "fd", "mqueue", ".lxc", ".lxd-mounts", ".udev":
-				continue
-			default:
-				var cp string
-				if containerPath != "" {
-					cp = filepath.Join(containerPath, filepath.Base(f.Name()))
-				}
-				sub, err := getDevices(filepath.Join(path, f.Name()), cp)
-				if err != nil {
-					return nil, err
-				}
-
-				out = append(out, sub...)
-				continue
-			}
-		case f.Name() == "console":
-			continue
-		}
-		device, err := deviceFromPath(filepath.Join(path, f.Name()))
-		if err != nil {
-			if err == errNotADevice {
-				continue
-			}
-			if os.IsNotExist(err) {
-				continue
-			}
-			return nil, err
-		}
-		if containerPath != "" {
-			device.Path = filepath.Join(containerPath, filepath.Base(f.Name()))
-		}
-		out = append(out, *device)
-	}
-	return out, nil
-}
-
-func deviceFromPath(path string) (*specs.LinuxDevice, error) {
-	var stat unix.Stat_t
-	if err := unix.Lstat(path, &stat); err != nil {
-		return nil, err
-	}
-
-	var (
-		devNumber = uint64(stat.Rdev)
-		major     = unix.Major(devNumber)
-		minor     = unix.Minor(devNumber)
-	)
-	if major == 0 {
-		return nil, errNotADevice
-	}
-
-	var (
-		devType string
-		mode    = stat.Mode
-	)
-	switch {
-	case mode&unix.S_IFBLK == unix.S_IFBLK:
-		devType = "b"
-	case mode&unix.S_IFCHR == unix.S_IFCHR:
-		devType = "c"
-	}
-	fm := os.FileMode(mode &^ unix.S_IFMT)
-	return &specs.LinuxDevice{
-		Type:     devType,
-		Path:     path,
-		Major:    int64(major),
-		Minor:    int64(minor),
-		FileMode: &fm,
-		UID:      &stat.Uid,
-		GID:      &stat.Gid,
-	}, nil
 }
 
 // WithCPUCFS sets the container's Completely fair scheduling (CFS) quota and period
